Load 3.3.157 into branch/prcs-import-33.
git-svn-id: svn://svn.bearnip.com/ldmud/branch/prcs-import-33@369 56160b6e-f8f8-0310-bd54-bccebfe6a2f2

+++ src/simulate.c
@@ -1344,7 +1344,8 @@ load_object_error(const char *msg, const char *name, namechain_t *chain)
     /* Make a local copy of the message so as not to leak memory */
     buf = alloca(strlen(sbuf.buf)+1);
     if (!buf)
-        error("Out of stack memory (%lu bytes)\n", strlen(sbuf.buf)+1);
+        error("Out of stack memory (%lu bytes)\n"
+             , (unsigned long) strlen(sbuf.buf)+1);
     strcpy(buf, sbuf.buf);
     strbuf_free(&sbuf);
